requirements.txt
babel==2.18.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
courlan==1.3.2
cssselect==1.4.0
dateparser==1.4.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.25.2
Flask==3.1.3
htmldate==1.9.4
idna==3.11
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.3
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.4
MarkupSafe==3.0.3
newspaper3k==0.2.8
nltk==3.9.4
pillow==12.1.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.3.32
requests==2.33.0
requests-file==3.0.1
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.3
tinysegmenter==0.3
tld==0.13.2
tldextract==5.3.1
tqdm==4.67.3
trafilatura==2.0.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.7
